mcp/requirements.txt
mcp>=1.0.0
paramiko>=3.0.0
pyyaml>=6.0
